feat: ➕ add getViews
