+++ packages/blog2/pages/api/views.ts
@@ -2,6 +2,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { FIREBASE_VIEWS_REF } from "../../src/constants/FIREBASE_VIEWS_REF";
 import { firebaseDb } from "../../src/functions/firebaseDb";
+import { getViews } from "../../src/functions/getViews";
 import { ViewsApi } from "../../src/types/ViewsApi";
 
 interface DataSnapshot {
@@ -37,13 +38,7 @@ const views = async (
   }
 
   if (request.method === "GET") {
-    // you cannot use destructuring here
-    const snapshot = await firebaseDb
-      .ref(FIREBASE_VIEWS_REF)
-      .child(slug)
-      .once("value");
-
-    const views = snapshot.val();
+    const { views } = await getViews({ slug });
 
     return response.status(200).json({ views });
   }
